Implement routing
We've defined two routes: one for the top recent scorers and one for the
top all time scorers. We added a nice list of links in the navbar to
view whichever.

I realized that the app component has no need for the top 100 service at
all. Now it's the camper table component that directly consumes that
service.

Speaking of the service, I promisified its API.