Highlight currently active route in the navbar
This one's amazing! `routerLinkActive` doesn't just apply to the actual
`<a>` element with the `routerLink`. It seems it can also be applied to
its parent element.

Here, Bootstrap has defined an `active` class to the `<li>`s that
enclose `<a>`s in a navbar. We added `routerLinkActive` to those
`<li>`s, and the result simply amazes me.