Clear table on route change
When switching from one route to the next, the table was still visible
for some time until the new response arrives. I find it a bit
disorienting, since the navbar indicates that you've selected a new
route but the old data is still there.

Now the table's wiped out while waiting for the new response.