Handle erroneous routes
Here we distinguish between `campers` having a value, being `null`, and
being `undefined`. The loading message is shown when the array is
undefined. Meanwhile we tell users there's nothing in the page if the
array is null. We only set the array to null when the user entered an
invalid URL (like, say, `/recentz`)